requirements.txt
streamlit
transformers
pillow
torch
pytesseract
